benerin  Solid payment

+++ src/main/java/com/app/smartdrive/api/dto/payment/Request/UserAccounts/UserAccountsDtoRequests.java
@@ -4,10 +4,12 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UserAccountsDtoRequests  {
     private String noRekening;
     @NotNull(message = "Tidak Boleh Kosong")
-    private Double nominall;
+    private BigDecimal nominall;
     
 }
